some changes

+++ src/cellsociety/GameSimulation/DeadCell.java
@@ -18,7 +18,7 @@ public class DeadCell extends Cell {
     }
 
     @Override
-    public void updateCell(GridEntry entry, Set<GridEntry> emptyCells, int[] parameters) {
+    public void updateCell(GridEntry entry, Set<GridEntry> emptyCells, double[] parameters) {
         Boolean enoughNeighbors = checkNeighbors(entry);
         if(enoughNeighbors){
             Cell liveCell = new LiveCell(entry);
