+++ src/cellsociety/PercolationSimulation/WaterCell.java
@@ -17,7 +17,7 @@ public class WaterCell extends Cell {
     }
 
     @Override
-    public void updateCell(GridEntry entry, Set<GridEntry> emptyCells, int[] parameters) {
+    public void updateCell(GridEntry entry, Set<GridEntry> emptyCells, double[] parameters) {
         entry.setNextStepCell(this);
     }
